Revert ProjectRow method location
* revert ShowStars method location
