+++ apps/bestofjs-nextjs/src/app/api/og/og-utils.tsx
@@ -59,53 +59,3 @@ export function StarIcon() {
     </svg>
   );
 }
-
-export function ProjectRow({
-  project,
-  rank,
-}: {
-  project: BestOfJS.Project;
-  rank: number;
-}) {
-  return (
-    <Box
-      style={{
-        color: "white",
-        gap: 24,
-        alignItems: "center",
-        borderBottom: "1px",
-        borderColor: "#3d3d42",
-        borderStyle: "dashed",
-        borderTopWidth: rank === 1 ? 1 : 0,
-        padding: "8px 0",
-      }}
-    >
-      <Box style={{ color: mutedColor }}>#{rank}</Box>
-      <Box>
-        <ProjectLogo project={project} size={80} />
-      </Box>
-      <Box
-        style={{
-          justifyContent: "space-between",
-          flex: 1,
-        }}
-      >
-        <div>{project.name}</div>
-        <Box>
-          <ShowStars project={project} />
-        </Box>
-      </Box>
-    </Box>
-  );
-}
-
-function ShowStars({ project }: { project: BestOfJS.Project }) {
-  return (
-    <Box
-      style={{ flexDirection: "row", alignItems: "center", color: mutedColor }}
-    >
-      <Box>{`+${project.trends.daily}`}</Box>
-      <StarIcon />
-    </Box>
-  );
-}
